+++ packages/@headlessui-react/src/types.ts
@@ -1,7 +1,9 @@
 import { ReactNode, ReactElement } from 'react'
-// A unique placeholder we can use as some defaults. This is nice because we can use this instead of
+
+// A unique placeholder we can use as a default. This is nice because we can use this instead of
 // defaulting to null / never / ... and possibly collide with actual data.
-const __: unique symbol = Symbol('__placeholder__')
+// Ideally we use a unique symbol here.
+let __ = '1D45E01E-AF44-47C4-988A-19A94EBAF55C' as const
 export type __ = typeof __
 
 export type Expand<T> = T extends infer O ? { [K in keyof O]: O[K] } : never